+++ source/tests/SourcesTests.cpp
@@ -165,12 +165,12 @@ TEST_CASE("Star disc", "[sources]")
     int const BELT_NUMBER{ 10 };
     double f[BELT_NUMBER]{};
 
-    for (std::int64_t x=0; x!=200; ++x)
+    for (std::int32_t x=0; x!=200; ++x)
     {
-        for (std::int64_t y=0; y != 200; ++y)
+        for (std::int32_t y=0; y != 200; ++y)
         {
-            double dx = 0.01 * static_cast<double>(1.0 * x - 99.5);
-            double dy = 0.01 * static_cast<double>(1.0 * y - 99.5);
+            double dx = 0.01 * (1.0 * x - 99.5);
+            double dy = 0.01 * (1.0 * y - 99.5);
 
             double const r = std::sqrt(dx * dx + dy * dy);
